commit with changes for wsdl2rest generator
