--- app/templates/blueprint/README.md
@@ -1,19 +1,0 @@
-<%= userProps.appName %>
-=============
-
-Camel Project for Blueprint 
-=========================================
-To build this project use
-```bash
-    mvn install
-```
-
-To run the project you can execute the following Maven goal
-```bash
-    mvn camel:run
-```
-
-For more help see the Apache Camel documentation
-
-    http://camel.apache.org/
-
