--- app/templates/blueprint/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-    Copyright 2014-2017, Red Hat, Inc. and/or its affiliates, and individual
-    contributors by the @authors tag. 
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
--->
-<!-- Configures the Camel Context-->
-<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.osgi.org/xmlns/blueprint/v1.0.0 https://www.osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd                            http://camel.apache.org/schema/blueprint http://camel.apache.org/schema/blueprint/camel-blueprint.xsd">
-    <camelContext id="_context1" xmlns="http://camel.apache.org/schema/blueprint">
-        <route id="_route1">
-            <from uri="timer://simpleTimer?period=1000"/>
-            <setBody>
-                <simple>Hello from timer at ${header.firedTime}></simple>
-            </setBody>
-            <to uri="stream:out"/>
-        </route>
-    </camelContext>
-</beans>
