--- app/templates/spring/pom.xml
@@ -1,172 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-	<modelVersion>4.0.0</modelVersion>
-
-	<groupId><%= userProps.package %></groupId>
-	<artifactId><%= userProps.name %></artifactId>
-	<version>1.0.0-SNAPSHOT</version>
-	<packaging>jar</packaging>
-
-	<name><%= userProps.name %></name>
-
-	<licenses>
-		<license>
-			<name>Apache License, Version 2.0</name>
-			<distribution>repo</distribution>
-			<url>http://www.apache.org/licenses/LICENSE-2.0.html</url>
-		</license>
-	</licenses>
-
-	<properties>
-		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-		<version.maven-bundle-plugin>3.2.0</version.maven-bundle-plugin>
-		<camel.version><%= userProps.camelVersion %></camel.version>
-	</properties>
-
-	<dependencies>
-		<dependency>
-			<groupId>org.apache.camel</groupId>
-			<artifactId>camel-core</artifactId>
-      		<version>${camel.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.camel</groupId>
-			<artifactId>camel-spring</artifactId>
-      		<version>${camel.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.camel</groupId>
-			<artifactId>camel-stream</artifactId>
-      		<version>${camel.version}</version>
-		</dependency>
-
-		<!-- Testing & Camel Plugin -->
-		<dependency>
-			<groupId>org.apache.camel</groupId>
-			<artifactId>camel-test-spring</artifactId>
-      		<version>${camel.version}</version>
-		</dependency>
-	</dependencies>
-
-	<repositories>
-		<repository>
-			<id>fuse-public-repository</id>
-			<name>FuseSource Community Release Repository</name>
-			<url>https://repo.fusesource.com/nexus/content/groups/public</url>
-			<snapshots>
-				<enabled>false</enabled>
-			</snapshots>
-			<releases>
-				<enabled>true</enabled>
-				<updatePolicy>never</updatePolicy>
-			</releases>
-		</repository>
-		<repository>
-			<id>red-hat-ga-repository</id>
-			<name>Red Hat GA Repository</name>
-			<url>https://maven.repository.redhat.com/ga</url>
-			<releases>
-				<enabled>true</enabled>
-				<updatePolicy>never</updatePolicy>
-			</releases>
-			<snapshots>
-				<enabled>false</enabled>
-			</snapshots>
-		</repository>
-		<repository>
-			<releases>
-				<enabled>true</enabled>
-				<updatePolicy>never</updatePolicy>
-			</releases>
-			<snapshots>
-				<enabled>false</enabled>
-			</snapshots>
-			<id>red-hat-ea-repository</id>
-			<name>Red Hat EA Repository</name>
-			<url>https://maven.repository.redhat.com/earlyaccess/all</url>
-		</repository>
-		<!-- To remove when a long-term version will be available -->
-		<repository>
-			<releases>
-				<enabled>true</enabled>
-				<updatePolicy>never</updatePolicy>
-			</releases>
-			<snapshots>
-				<enabled>false</enabled>
-			</snapshots>
-			<id>jboss-ea-repository</id>
-			<name>Red Hat JBoss Early Access Repository</name>
-			<url>http://repository.jboss.org/nexus/content/groups/ea</url>
-		</repository>
-	</repositories>
-
-	<pluginRepositories>
-		<pluginRepository>
-			<id>fuse-public-repository</id>
-			<name>FuseSource Community Release Repository</name>
-			<url>https://repo.fusesource.com/nexus/content/groups/public</url>
-			<snapshots>
-				<enabled>false</enabled>
-			</snapshots>
-			<releases>
-				<enabled>true</enabled>
-				<updatePolicy>never</updatePolicy>
-			</releases>
-		</pluginRepository>
-		<pluginRepository>
-			<id>red-hat-ga-repository</id>
-			<name>Red Hat GA Repository</name>
-			<url>https://maven.repository.redhat.com/ga</url>
-			<releases>
-				<enabled>true</enabled>
-				<updatePolicy>never</updatePolicy>
-			</releases>
-			<snapshots>
-				<enabled>false</enabled>
-			</snapshots>
-		</pluginRepository>
-				<pluginRepository>
-			<releases>
-				<enabled>true</enabled>
-				<updatePolicy>never</updatePolicy>
-			</releases>
-			<snapshots>
-				<enabled>false</enabled>
-			</snapshots>
-			<id>red-hat-ea-repository</id>
-			<name>Red Hat EA Repository</name>
-			<url>https://maven.repository.redhat.com/earlyaccess/all</url>
-		</pluginRepository>
-		<!-- To remove when a long-term version will be available -->
-		<pluginRepository>
-			<releases>
-				<enabled>true</enabled>
-				<updatePolicy>never</updatePolicy>
-			</releases>
-			<snapshots>
-				<enabled>false</enabled>
-			</snapshots>
-			<id>jboss-ea-repository</id>
-			<name>Red Hat JBoss Early Access Repository</name>
-			<url>http://repository.jboss.org/nexus/content/groups/ea</url>
-		</pluginRepository>
-	</pluginRepositories>
-
-	<build>
-		<plugins>
-		<!-- Allows the routes to be run via 'mvn camel:run' -->
-		<plugin>
-			<groupId>org.apache.camel</groupId>
-			<artifactId>camel-maven-plugin</artifactId>
-				<configuration>
-					<applicationContextUri>META-INF/spring/camel-context.xml</applicationContextUri>
-				</configuration>
-			</plugin>
-		</plugins>
-	</build>
-
-	<description>Empty Camel Spring Example</description>
-</project>
