+++ generators/wsdl2rest/dependency-reduced-pom.xml
@@ -1,0 +1,64 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <groupId>org.fusesource.yeoman.wsdl2rest</groupId>
+  <artifactId>org.fusesource.ide.wsdl2rest</artifactId>
+  <name>Red Hat Fuse Tooling :: Camel Editor :: Plugins :: Wsdl2Rest Implementation</name>
+  <version>0.1.0-SNAPSHOT</version>
+  <build>
+    <finalName>wsdl2rest</finalName>
+    <plugins>
+      <plugin>
+        <artifactId>maven-clean-plugin</artifactId>
+        <configuration>
+          <filesets>
+            <fileset>
+              <directory>${basedir}/lib</directory>
+            </fileset>
+          </filesets>
+        </configuration>
+      </plugin>
+      <plugin>
+        <artifactId>maven-shade-plugin</artifactId>
+        <version>3.1.1</version>
+        <executions>
+          <execution>
+            <phase>package</phase>
+            <goals>
+              <goal>shade</goal>
+            </goals>
+            <configuration>
+              <transformers>
+                <transformer>
+                  <mainClass>org.jboss.fuse.wsdl2rest.impl.Wsdl2Rest</mainClass>
+                </transformer>
+              </transformers>
+            </configuration>
+          </execution>
+        </executions>
+      </plugin>
+    </plugins>
+  </build>
+  <dependencyManagement>
+    <dependencies>
+      <dependency>
+        <groupId>org.apache.camel</groupId>
+        <artifactId>camel-parent</artifactId>
+        <version>${version.apache.camel}</version>
+        <type>pom</type>
+        <scope>import</scope>
+      </dependency>
+    </dependencies>
+  </dependencyManagement>
+  <properties>
+    <version.javax.ws.rs.api>2.1.0.redhat-1</version.javax.ws.rs.api>
+    <version.wsdl4j>1.6.3.redhat-2</version.wsdl4j>
+    <version.args4j>2.33</version.args4j>
+    <version.apache.camel>2.21.0.fuse-000116</version.apache.camel>
+    <version.javaparser>2.5.1</version.javaparser>
+    <version.apache.cxf>3.1.11.fuse-000285</version.apache.cxf>
+    <version.wsdl2rest>0.7.0.fuse-000086-redhat-3</version.wsdl2rest>
+    <version.junit>4.12.0.redhat-003</version.junit>
+    <version.apache.velocity>1.7.0.redhat-5</version.apache.velocity>
+  </properties>
+</project>
