+++ generators/wsdl2rest/pom.xml.individual-jars
@@ -1,0 +1,168 @@
+<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" 
+    xmlns="http://maven.apache.org/POM/4.0.0" 
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>org.fusesource.ide.wsdl2rest</artifactId>
+    <!-- <packaging>eclipse-plugin</packaging> -->
+    <name>Red Hat Fuse Tooling :: Camel Editor :: Plugins :: Wsdl2Rest Implementation</name>
+    <version>0.1.0-SNAPSHOT</version>
+    <groupId>org.fusesource.yeoman.wsdl2rest</groupId>
+
+    <properties>
+        <version.apache.camel>2.20.1</version.apache.camel>
+        <version.apache.cxf>3.2.1</version.apache.cxf>
+        <version.apache.velocity>1.7</version.apache.velocity>
+        <version.eclipse.jetty>9.4.6.v20170531</version.eclipse.jetty>
+        <version.args4j>2.33</version.args4j>
+        <version.javaparser>2.5.1</version.javaparser>
+        <version.javax.ws.rs.api>2.1</version.javax.ws.rs.api>
+        <version.junit>4.12</version.junit>
+        <version.wsdl4j>1.6.3</version.wsdl4j>
+        <version.wsdl2rest>0.8.0.fuse-720006</version.wsdl2rest>
+    </properties>
+
+    <dependencyManagement>
+        <dependencies>
+            <dependency>
+                <groupId>org.apache.camel</groupId>
+                <artifactId>camel-parent</artifactId>
+                <version>${version.apache.camel}</version>
+                <scope>import</scope>
+                <type>pom</type>
+            </dependency>
+        </dependencies>
+    </dependencyManagement>
+
+    <dependencies>
+        <dependency>
+            <groupId>args4j</groupId>
+            <artifactId>args4j</artifactId>
+            <version>${version.args4j}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.github.javaparser</groupId>
+            <artifactId>javaparser-core</artifactId>
+            <version>${version.javaparser}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.camel</groupId>
+            <artifactId>camel-spring</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.cxf</groupId>
+            <artifactId>cxf-codegen-plugin</artifactId>
+            <version>${version.apache.cxf}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.cxf</groupId>
+            <artifactId>cxf-rt-rs-extension-providers</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.cxf</groupId>
+            <artifactId>cxf-rt-transports-http-jetty</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.velocity</groupId>
+            <artifactId>velocity</artifactId>
+            <version>${version.apache.velocity}</version>
+        </dependency>
+        <dependency>
+            <groupId>wsdl4j</groupId>
+            <artifactId>wsdl4j</artifactId>
+            <version>${version.wsdl4j}</version>
+        </dependency>
+        <dependency>
+            <groupId>javax.ws.rs</groupId>
+            <artifactId>javax.ws.rs-api</artifactId>
+            <version>${version.javax.ws.rs.api}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-log4j12</artifactId>
+            <!-- Add in scope compile so it ends up in the manifest Class-Path -->
+        </dependency>
+        <dependency>
+            <groupId>log4j</groupId>
+            <artifactId>log4j</artifactId>
+        </dependency>
+    </dependencies>
+
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-dependency-plugin</artifactId>
+                <executions>
+                    <execution>
+                        <id>copy-wsdl2rest</id>
+                        <phase>generate-resources</phase>
+                        <goals>
+                            <goal>copy</goal>
+                        </goals>
+                        <configuration>
+                            <stripVersion>true</stripVersion>
+                            <outputDirectory>${basedir}/lib/</outputDirectory>
+                            <skip>false</skip>
+                            <artifactItems>
+                                <!-- Impl -->
+                                <artifactItem>
+                                    <groupId>org.jboss.fuse.wsdl2rest</groupId>
+                                    <artifactId>wsdl2rest-impl</artifactId>
+                                    <version>${version.wsdl2rest}</version>
+                                </artifactItem>
+                                <!-- Util -->
+                                <artifactItem>
+                                    <groupId>org.jboss.fuse.wsdl2rest</groupId>
+                                    <artifactId>wsdl2rest-util</artifactId>
+                                    <version>${version.wsdl2rest}</version>
+                                </artifactItem>
+                            </artifactItems>
+                        </configuration>
+                    </execution>
+                    <execution>
+                        <id>copy-dependencies</id>
+                        <phase>generate-sources</phase>
+                        <goals>
+                            <goal>copy-dependencies</goal>
+                        </goals>
+                        <configuration>
+                            <excludeScope>system</excludeScope>
+                            <!-- <stripVersion>true</stripVersion> -->
+                            <outputDirectory>${basedir}/lib/</outputDirectory>
+                            <skip>false</skip>
+                        </configuration>
+                    </execution>
+                </executions>
+            </plugin>
+            <plugin>
+                <!-- make sure lib dir is removed after clean to avoid "dirty" build -->
+                <artifactId>maven-clean-plugin</artifactId>
+                <configuration>
+                    <filesets>
+                        <fileset>
+                            <directory>${basedir}/lib</directory>
+                        </fileset>
+                    </filesets>
+                </configuration>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-enforcer-plugin</artifactId>
+                <executions>
+                    <execution>
+                        <id>ban-uberjars</id>
+                        <configuration>
+                            <skip>true</skip>
+                        </configuration>
+                    </execution>
+                </executions>
+            </plugin>
+        </plugins>
+    </build>
+
+</project>
