+++ generators/wsdl2rest/pom.xml.jar-with-dependencies
@@ -1,0 +1,155 @@
+<project
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
+	xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+	<modelVersion>4.0.0</modelVersion>
+
+	<artifactId>org.fusesource.ide.wsdl2rest</artifactId>
+	<!-- <packaging>eclipse-plugin</packaging> -->
+	<name>Red Hat Fuse Tooling :: Camel Editor :: Plugins :: Wsdl2Rest Implementation</name>
+	<version>0.1.0-SNAPSHOT</version>
+	<groupId>org.fusesource.yeoman.wsdl2rest</groupId>
+
+	<properties>
+        <version.apache.camel>2.21.0.fuse-000116</version.apache.camel>
+        <version.apache.cxf>3.1.11.fuse-000285</version.apache.cxf>
+        <version.apache.velocity>1.7.0.redhat-5</version.apache.velocity>
+        <version.args4j>2.33</version.args4j>
+        <version.javaparser>2.5.1</version.javaparser>
+        <version.javax.ws.rs.api>2.1.0.redhat-1</version.javax.ws.rs.api>
+        <version.junit>4.12.0.redhat-003</version.junit>
+        <version.wsdl4j>1.6.3.redhat-2</version.wsdl4j>
+		<version.wsdl2rest>0.7.0.fuse-000086-redhat-3</version.wsdl2rest>
+	</properties>
+
+	<dependencyManagement>
+		<dependencies>
+			<dependency>
+				<groupId>org.apache.camel</groupId>
+				<artifactId>camel-parent</artifactId>
+				<version>${version.apache.camel}</version>
+				<scope>import</scope>
+				<type>pom</type>
+			</dependency>
+		</dependencies>
+    </dependencyManagement>
+
+    <dependencies>
+        <dependency>
+            <groupId>args4j</groupId>
+            <artifactId>args4j</artifactId>
+            <version>${version.args4j}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.github.javaparser</groupId>
+            <artifactId>javaparser-core</artifactId>
+            <version>${version.javaparser}</version>
+        </dependency>
+        <dependency>
+            <groupId>javax.ws.rs</groupId>
+            <artifactId>javax.ws.rs-api</artifactId>
+            <version>${version.javax.ws.rs.api}</version>
+        </dependency>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <version>${version.junit}</version>
+        </dependency>
+		<dependency>
+			<groupId>org.apache.camel</groupId>
+			<artifactId>camel-spring</artifactId>
+            <version>${version.apache.camel}</version>
+        </dependency>        
+        <dependency>
+            <groupId>org.apache.cxf</groupId>
+            <artifactId>cxf-codegen-plugin</artifactId>
+            <version>${version.apache.cxf}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.cxf</groupId>
+            <artifactId>cxf-tools-wsdlto-frontend-jaxws</artifactId>
+            <version>${version.apache.cxf}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.velocity</groupId>
+            <artifactId>velocity</artifactId>
+            <version>${version.apache.velocity}</version>
+        </dependency>
+        <dependency>
+            <groupId>wsdl4j</groupId>
+            <artifactId>wsdl4j</artifactId>
+            <version>${version.wsdl4j}</version>
+        </dependency>
+		<dependency>
+			<groupId>org.slf4j</groupId>
+			<artifactId>slf4j-api</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.slf4j</groupId>
+			<artifactId>slf4j-log4j12</artifactId>
+			<!-- Add in scope compile so it ends up in the manifest Class-Path -->
+		</dependency>
+		<dependency>
+			<groupId>log4j</groupId>
+			<artifactId>log4j</artifactId>
+        </dependency>        
+		<dependency>
+			<groupId>org.jboss.fuse.wsdl2rest</groupId>
+			<artifactId>wsdl2rest-impl</artifactId>
+			<version>${version.wsdl2rest}</version>
+		</dependency>
+		<!-- Util -->
+		<dependency>
+			<groupId>org.jboss.fuse.wsdl2rest</groupId>
+			<artifactId>wsdl2rest-util</artifactId>
+			<version>${version.wsdl2rest}</version>
+		</dependency>
+    </dependencies>
+
+	<build>
+		<finalName>wsdl2rest</finalName>
+		<plugins>
+			<plugin>
+				<!-- make sure lib dir is removed after clean to avoid "dirty" build -->
+				<artifactId>maven-clean-plugin</artifactId>
+				<configuration>
+					<filesets>
+						<fileset>
+							<directory>${basedir}/lib</directory>
+						</fileset>
+					</filesets>
+				</configuration>
+			</plugin>
+
+			<!-- Maven Assembly Plugin -->
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-assembly-plugin</artifactId>
+				<version>2.4.1</version>
+				<configuration>
+					<!-- get all project dependencies -->
+					<descriptorRefs>
+						<descriptorRef>jar-with-dependencies</descriptorRef>
+					</descriptorRefs>
+					<!-- MainClass in mainfest make a executable jar -->
+					<archive>
+					  <manifest>
+						<mainClass>org.jboss.fuse.wsdl2rest.impl.Wsdl2Rest</mainClass>
+					  </manifest>
+					</archive>
+					<outputDirectory>${basedir}/lib/</outputDirectory>
+				</configuration>
+				<executions>
+				  <execution>
+					<id>make-assembly</id>
+                    <!-- bind to the packaging phase -->
+					<phase>package</phase> 
+					<goals>
+						<goal>single</goal>
+					</goals>
+				  </execution>
+				</executions>
+			</plugin>
+		</plugins>
+	</build>
+
+</project>
